src/test/java/toDoProject/toDo/service/MemberServiceTest.java
package toDoProject.toDo.service;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.data.domain.Example;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.data.repository.query.FluentQuery;
import toDoProject.toDo.domain.Member;
import toDoProject.toDo.dto.MemberDTO;
import toDoProject.toDo.repository.MemberRepository;

import java.util.List;
import java.util.Optional;
import java.util.function.Function;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class MemberServiceTest {

    @Autowired
    MemberService memberService;

    @Test
    void findOne() {
        Member member = new Member("kim", "zzq", "3902");
        Long id = memberService.create(member);

        assertThat(memberService.findOne(id).getUsername()).isEqualTo(member.getUsername());
    }
}
